feat: userId 자동 바인딩 연동

+++ src/main/java/com/idle/weather/level/repository/LevelJpaRepository.java
@@ -2,5 +2,5 @@ package com.idle.weather.level.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LevelJpaRepository extends JpaRepository<LevelEntity , Long> {
+public interface LevelJpaRepository extends JpaRepository<LevelEntity , Integer> {
 }
